requirements2.txt
awsebcli==3.8.8
blessed==1.9.5
botocore==1.4.87
cement==2.8.2
colorama==0.3.7
Django==1.10.3
django-appconf==1.0.2
django-colorfield==0.1.10
django-extensions==1.7.4
django-imagekit==3.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
jmespath==0.9.0
pathspec==0.5.0
pilkit==1.1.13
Pillow==3.4.2
pkg-resources==0.0.0
python-dateutil==2.6.0
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.10.0
tabulate==0.7.5
wcwidth==0.1.7
websocket-client==0.40.0
